Django/requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-countries==5.5
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2